corrigido score para fazer operações com doubles e retornar um double

+++ java/src/score/MDLScore.java
@@ -7,8 +7,8 @@ import bayessian.RandomVariable;
 public class MDLScore extends LLScore {
 	
 	@Override
-	public int getScore(BayessianNetwork<? extends RandomVariable> bayessian, Dataset dataset) {
-		int score = 0;
+	public double getScore(BayessianNetwork<? extends RandomVariable> bayessian, Dataset dataset) {
+		double score = 0;
 		//score = super.getScore() - log(N)*getNetworkComplexity()/2.0;
 		return score;
 	}
